% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_mat.R
\name{corr_mat}
\alias{corr_mat}
\title{Correlation Matrix of Numeric NBA Stats in a Given Year}
\usage{
corr_mat(year)
}
\arguments{
\item{Year}{Which NBA season}
}
\description{
This function outputs the correlation matrix of numeric NBA stats for a given year
}
\examples{
corr_mat()
}
\keyword{NBA}
\keyword{correlation}
\keyword{matrix}
\keyword{numeric}
\keyword{stats}
