% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_nba.R
\name{corr_nba}
\alias{corr_nba}
\title{Return Season Correlation Matrix NBA Dataset}
\usage{
corr_nba(year_in)
}
\arguments{
\item{year_in}{Season to be evaluated.}
}
\description{
This function returns the correlation matrix for all the variables included in the
NBA dataset for a given NBA season.
}
\examples{
corr_nba(1991)
}
\keyword{correlation}
\keyword{matrix.}
\keyword{season,}
