% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\name{correlation_matrix}
\alias{correlation_matrix}
\title{An nbapackage function}
\usage{
correlation_matrix(year)
}
\arguments{
\item{year}{The year from which you want the oldest player.}
}
\description{
This function will give out the correlation matrix of all continuous variables from a given year/season
}
\examples{
correlation_matrix()
}
\keyword{correlation}
\keyword{matrix}
