% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrNumerics.R
\name{corrNumerics}
\alias{corrNumerics}
\title{A Function}
\usage{
corrNumerics(yr)
}
\arguments{
\item{year}{an integer, which is the year}
}
\description{
This function takes an individual year as input, and output a correlation matrix for all numeric variables in the NBA data for that year.
}
\examples{
corrNumerics(2000)
}
\keyword{corrNumerics}
