% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_points.R
\name{most_points}
\alias{most_points}
\title{Leading Scorer in a Given Season}
\usage{
most_points(year)
}
\arguments{
\item{Year}{Which NBA season}
}
\description{
This function outputs the leading scorer in the NBA in a given year and his corresponding point total.
}
\examples{
most_points()
}
\keyword{NBA}
\keyword{leading}
\keyword{scorer}
