% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NbaPackage.R
\docType{package}
\name{NbaPackage}
\alias{NbaPackage}
\alias{NbaPackage-package}
\title{NbaPackage: a package for computing useful statistics about NBA.}
\description{
The foo package provides three categories of important functions: 
return_oldest, return_avg_FTpctg and corr_nba.
}
