% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_pl.R
\name{oldest_pl}
\alias{oldest_pl}
\title{Oldest Player in a Given Season}
\usage{
oldest_pl(year)
}
\arguments{
\item{Year}{Which NBA season}
}
\description{
This function outputs the oldest player in the NBA in a given year and his corresponding age.
}
\examples{
oldest_pl()
}
\keyword{NBA}
\keyword{oldest}
\keyword{player}
