% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oldest_Player.R
\name{Oldest_Player}
\alias{Oldest_Player}
\title{A NBA Player Function}
\usage{
Oldest_Player(x)
}
\description{
This function allows you to select the oldest player in a specific year.
}
\examples{
Oldest_Player(1991)
}
\keyword{oldest}
\keyword{player}
