% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldestPlayer.R
\name{oldestPlayer}
\alias{oldestPlayer}
\title{A Function}
\usage{
oldestPlayer(yr)
}
\arguments{
\item{year}{an integer, which is the year of the season}
}
\description{
This function returns the name of oldest NBA player in a particular season.
}
\examples{
oldestPlayer(2000)
}
\keyword{oldestPlayer}
