% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_avg_FTpctg.R
\name{return_avg_FTpctg}
\alias{return_avg_FTpctg}
\title{Return Average FT Percentage}
\usage{
return_avg_FTpctg(player_in)
}
\arguments{
\item{player_in}{Player to be evaluated.}
}
\description{
This function returns the average FT percentage for a given player across
all the NBA seasons played by the player.
}
\examples{
player_in("Robert Parish")
}
\keyword{FT}
\keyword{player,}
