% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_oldest.R
\name{return_oldest}
\alias{return_oldest}
\title{Return Oldest NBA Player}
\usage{
return_oldest(year_in)
}
\arguments{
\item{year_in}{Season to be evaluated.}
}
\description{
This function returns the oldest NBA player in a given season.
}
\examples{
return_oldest(1991)
}
\keyword{player,}
\keyword{season}
