% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_scorer.R
\name{top_scorer}
\alias{top_scorer}
\title{An nbapackage function}
\usage{
top_scorer(year)
}
\arguments{
\item{year}{The year from which you want the top scorer.}
}
\description{
This function will give out the player who scored the most amount of points in the NBA for the selected year.
}
\examples{
top_scorer()
}
\keyword{scorer}
\keyword{top}
