% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearView.R
\name{yearView}
\alias{yearView}
\title{A Function}
\usage{
yearView(yr)
}
\arguments{
\item{year}{an integer, which is the year of the season}
}
\description{
This function returns the information about an NBA player and their performance in a particular season.
}
\examples{
yearView(2000)
}
\keyword{yearView}
