% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{An nbapackage function}
\usage{
oldest_player(year)
}
\arguments{
\item{year}{The year from which you want the oldest player.}
}
\description{
This function will give out the oldest player by age in any given year/season.
}
\examples{
oldest_player()
}
\keyword{oldest}
\keyword{player}
